[Fix] Refactored listener code to improve wand usage
Previously, this would not cancel the event if the player was not
clicking an container block. This would cause the item to be used
normally if clicking on regular blocks, which would sometimes have
undesirable effects. The listener code has been rewritten to always
cancel the event, if the player is holding the wand. Then it checks for
a container block, and fills it if needed. This code also is smaller.