New features
Item names now have tab completion and are verified to avoid a crash. Item data via command line argument is no longer supported as it was done via numerical value and is also not very useful. The name of the command was also changed because it conflicts with other commands but the alias also makes it easy to use again.